+++ NEWS.md
@@ -4,3 +4,4 @@
 * Upgraded D3 to V5.
 * Moves the vizdraws js library to its own folder to make the development of a python package easier.
 * Added `percentage` as an option to set the x axis tick format to percentage.
+* Added `tense` as an option to use future or past tense in the description.